stop testing for valid connections when reading database.yml
first, this was not threadsafe. it was swapping the db connection on
ActiveRecord::Base, which is maintained by ActiveRecord as a class
singleton (not a thread var).

second, if this actually worked, it could automatically remove all
replicas and cause a stampede on master. we've consciously decided
that database config changes should be handled by the provisioning
system (chef).