require host application to decide on reconnect strategy
the verify! command is implemented as `reconnect! unless active?`. the
effect here was to prevent a query timeout from leaving the connection
in a bad state. but in current versions of rails it results in another
error (NoMethodError: undefined method rollback) because the current
transaction has already been auto-closed by the reconnection.

suggested fix is `reconnect: true` in database.yml. this is supported
by activerecord and allows exceptions to cause an expected rollback
before the next query re-establishes the connection.